chore(api): move attestation API to common package
