+++ app/cli/internal/action/attestation_init.go
@@ -21,9 +21,9 @@ import (
 	"fmt"
 	"strconv"
 
-	clientAPI "github.com/chainloop-dev/chainloop/app/cli/api/attestation/v1"
 	pb "github.com/chainloop-dev/chainloop/app/controlplane/api/controlplane/v1"
 	"github.com/chainloop-dev/chainloop/internal/attestation/crafter"
+	clientAPI "github.com/chainloop-dev/chainloop/internal/attestation/crafter/api/attestation/v1"
 )
 
 type AttestationInitOpts struct {
