+++ internal/attestation/crafter/api/attestation/v1/crafting_state_validations.go
