+++ internal/attestation/crafter/materials/artifact_test.go
@@ -20,8 +20,8 @@ import (
 	"testing"
 
 	"code.cloudfoundry.org/bytefmt"
-	attestationApi "github.com/chainloop-dev/chainloop/app/cli/api/attestation/v1"
 	contractAPI "github.com/chainloop-dev/chainloop/app/controlplane/api/workflowcontract/v1"
+	attestationApi "github.com/chainloop-dev/chainloop/internal/attestation/crafter/api/attestation/v1"
 	"github.com/chainloop-dev/chainloop/internal/attestation/crafter/materials"
 	"github.com/chainloop-dev/chainloop/internal/casclient"
 	mUploader "github.com/chainloop-dev/chainloop/internal/casclient/mocks"
