+++ internal/attestation/crafter/materials/cyclonedxjson.go
@@ -21,8 +21,8 @@ import (
 	"os"
 
 	cdx "github.com/CycloneDX/cyclonedx-go"
-	api "github.com/chainloop-dev/chainloop/app/cli/api/attestation/v1"
 	schemaapi "github.com/chainloop-dev/chainloop/app/controlplane/api/workflowcontract/v1"
+	api "github.com/chainloop-dev/chainloop/internal/attestation/crafter/api/attestation/v1"
 	"github.com/chainloop-dev/chainloop/internal/casclient"
 	"github.com/rs/zerolog"
 )
