+++ internal/attestation/crafter/materials/junit_xml.go
@@ -23,8 +23,8 @@ import (
 
 	"encoding/xml"
 
-	api "github.com/chainloop-dev/chainloop/app/cli/api/attestation/v1"
 	schemaapi "github.com/chainloop-dev/chainloop/app/controlplane/api/workflowcontract/v1"
+	api "github.com/chainloop-dev/chainloop/internal/attestation/crafter/api/attestation/v1"
 	"github.com/chainloop-dev/chainloop/internal/casclient"
 	junit "github.com/joshdk/go-junit"
 	"github.com/rs/zerolog"
