+++ internal/attestation/crafter/materials/materials_test.go
@@ -20,8 +20,8 @@ import (
 	"testing"
 	"time"
 
-	attestationApi "github.com/chainloop-dev/chainloop/app/cli/api/attestation/v1"
 	contractAPI "github.com/chainloop-dev/chainloop/app/controlplane/api/workflowcontract/v1"
+	attestationApi "github.com/chainloop-dev/chainloop/internal/attestation/crafter/api/attestation/v1"
 	"github.com/chainloop-dev/chainloop/internal/attestation/crafter/materials"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
