+++ internal/attestation/crafter/materials/openvex.go
@@ -20,8 +20,8 @@ import (
 	"fmt"
 	"os"
 
-	api "github.com/chainloop-dev/chainloop/app/cli/api/attestation/v1"
 	schemaapi "github.com/chainloop-dev/chainloop/app/controlplane/api/workflowcontract/v1"
+	api "github.com/chainloop-dev/chainloop/internal/attestation/crafter/api/attestation/v1"
 	"github.com/chainloop-dev/chainloop/internal/casclient"
 	"github.com/openvex/go-vex/pkg/vex"
 	"github.com/rs/zerolog"
