+++ internal/attestation/crafter/materials/sarif.go
@@ -19,8 +19,8 @@ import (
 	"context"
 	"fmt"
 
-	api "github.com/chainloop-dev/chainloop/app/cli/api/attestation/v1"
 	schemaapi "github.com/chainloop-dev/chainloop/app/controlplane/api/workflowcontract/v1"
+	api "github.com/chainloop-dev/chainloop/internal/attestation/crafter/api/attestation/v1"
 	"github.com/chainloop-dev/chainloop/internal/casclient"
 	"github.com/owenrumney/go-sarif/sarif"
 	"github.com/rs/zerolog"
